// src/app/state/movies/movies.types.ts
export enum MoviesActionTypes {
    LOAD_ALL_MOVIES = '[MOVIES] load all movies',
    LOAD_ALL_MOVIES_SUCCESS = '[MOVIES] load all movies successfully',
    LOAD_ALL_MOVIES_ERROR = '[MOVIES] load all movies error',

    LOAD_SINGLE_MOVIE = '[MOVIES] load single movie',
    LOAD_SINGLE_MOVIE_SUCCESS = '[MOVIES] load single movie successfully',
    LOAD_SINGLE_MOVIE_ERROR = '[MOVIES] load single movie error',

    SET_FAVOURATE_MOVIE = '[MOVIES] set favourate movie',
    SET_FAVOURATE_MOVIE_SUCCESS = '[MOVIES] set favourate movie successfully',
    SET_FAVOURATE_MOVIE_ERROR = '[MOVIES] set favourate movie error',
}